Stationary Distribution
For some reason, when I put this code into a function and included it in my main file, the function calculated the wrong distribution. Therefore, I just moved the code to calculate the distribution to its own file and included it in my compute file. Everything is running as it should now!