While loop working!
Finally got a working, accurate while loop! The reason the other loops were giving me the wrong aggregate capital and labor was the fact that I wasn't re-initializing the population distribution before re-solving the model. Thus, it was adding the various distributions to one another and getting an increasing aggregate labor and capital.